+++ src/InvestPool/PoolToken.sol
@@ -5,7 +5,7 @@ import {ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 
 contract PoolToken is ERC20("loan token", "lnt"), Ownable {
-    function mint(uint amount) external onlyOwner {
+    function mint(uint256 amount) external onlyOwner {
         _mint(msg.sender, amount);
     }
 }
